Refactor multiplex protocol module into dedicated submodules
