+++ crates/protocol/src/multiplex/mod.rs
@@ -1,0 +1,11 @@
+mod borrowed;
+mod frame;
+mod helpers;
+mod io;
+
+#[cfg(test)]
+mod tests;
+
+pub use borrowed::{BorrowedMessageFrame, BorrowedMessageFrames};
+pub use frame::MessageFrame;
+pub use io::{recv_msg, recv_msg_into, send_frame, send_msg};
